% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_yf_data.R
\name{clean_yf_data}
\alias{clean_yf_data}
\title{Clean Historical Yahoo Finance data}
\usage{
clean_yf_data(final_df, set_column_names, wd, RDSfilename)
}
\arguments{
\item{final_df}{The name of the object that was previously parsed. 
This will also be the name of the data frame to be created as output.}

\item{set_column_names}{Each column heading prefix (format: must use 
quotes).  For example, if this is set to "vix", then the column 
headings of the data frame will be "vix_open", "vix_high", "vix_low", 
"vix_close", and "vix_adjclose".}

\item{wd}{The path to the working directory (format: must use quotes) 
where the output RDS file will be saved.}

\item{RDSfilename}{The name (format: must use quotes) of the output RDS file.}
}
\description{
This function cleans the historical Yahoo Finance quote data
that was parsed using "parse_yf()".  The output is saved as a data frame 
and RDS file.
Please follow the steps in the vignette to understand at what 
point to use this function.  PLEASE NOTE THAT THIS FUNCTION WAS 
CREATED ON 2017-11-22.  THEREFORE, THIS FUNCTION MAY BECOME
OBSOLETE IF YAHOO FINANCE CHANGES THE WEBSITE CONFIGURATION AFTER 
2017-11-22 AS THE INPUTS USED IN THIS FUNCTION WERE CREATED 
BY FUNCTIONS THAT USE WEB SCRAPPING TECHNIQUES.
}
\examples{
\dontrun{
clean_yf_data(final_df = vix_data,
set_column_names = "vix",
wd = "/folder_name",
RDSfilename = "VIX"
)}

}
\author{
Stephanie Langeland stephanie.langeland@gmail.com
}
\keyword{EconFiData}
