% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Federal_Funds_Target_Rate}
\alias{Federal_Funds_Target_Rate}
\title{Federal_Funds_Target_Rate.rda}
\format{Federal_Funds_Target_Rate is a list of 13.  Access the data in the list using 
as.data.frame(Federal_Funds_Target_Rate$observations), which has 5,098 observations (rows)
and 4 variables: "realtime_start" and "realtime_end" which is the date the API call was made; 
"date" which is the date of each observations; and "value" which is the Federal Funds Target Rate.}
\source{
https://fred.stlouisfed.org/series/DFEDTAR
}
\usage{
{Federal_Funds_Target_Rate}
}
\description{
The documentation for the get_series_obs function includes
an example that obtains daily Federal Funds Target Rate data from 
1995-01-01 to 2008-12-15 using a FRED API key (details: 
https://research.stlouisfed.org/docs/api/api_key.html).  This is the data 
frame that was created.
}
\keyword{datasets}
