% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meeting_dates}
\alias{meeting_dates}
\title{meeting_dates.rda}
\format{meeting_dates is a data frame with 167 observations (rows) and 1 variable (column) 
named historical_text.  These data were scraped from the FED's website.}
\source{
The data from the following websites were scrapped on 2017-11-22:
"https://www.federalreserve.gov/monetarypolicy/fomchistorical1995.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical1996.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical1997.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical1998.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical1999.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2000.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2001.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2002.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2003.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2004.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2005.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2006.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2007.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2008.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2009.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2010.htm"
"https://www.federalreserve.gov/monetarypolicy/fomchistorical2011.htm"
}
\usage{
{meeting_dates}
}
\description{
The documentation for the get_historical_dates function includes
example #2 which collects FOMC meeting dates from 1995 - 2011.  This is the
data frame that was created.
}
\keyword{datasets}
