% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yf.R
\name{parse_yf}
\alias{parse_yf}
\title{Parse Yahoo Finance Data}
\usage{
parse_yf(set_page_source)
}
\arguments{
\item{set_page_source}{The object name of the page source obtained
by using the RSelenium function "getPageSource()".  There is no
default for this parameter, the user must take the preceeding
steps outlined in the vignette to understand how to create this
object.}
}
\description{
This function parses the data obtained from web scrapping 
historical quote Yahoo Finance data.  The output is saved in a data frame. 
Please follow the steps in the vignette to understand at what 
point to use this function.  PLEASE NOTE THAT THIS FUNCTION WAS 
CREATED ON 2017-11-22.  THEREFORE, THIS FUNCTION MAY BECOME
OBSOLETE IF YAHOO FINANCE CHANGES THE WEBSITE CONFIGURATION AFTER 
2017-11-22 AS THIS FUNCTION USES WEB SCRAPPING TECHNIQUES.
}
\examples{
\dontrun{
parse_yf(set_page_source = page_source)
}

}
\author{
Stephanie Langeland stephanie.langeland@gmail.com
}
\keyword{EconFiData}
