% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vix_data}
\alias{vix_data}
\title{vix_data.rda}
\format{vix_data is a data frame with 5,763 VIX observations (rows) and 6
variables (columns).  The variables are "vix_open": daily open value;
"vix_high": highest daily value; "vix_low": lowest daily value; 
"vix_close": daily close value; "vix_adjclose": daily adjusted close value.}
\source{
https://finance.yahoo.com/quote/%5EVIX/history?period1=789973200&period2=1512190800&interval=1d&filter=history&frequency=1d
}
\usage{
{vix_data.}
}
\description{
The documentation for the clean_yf_data function includes
an example that produces clean historical VIX data.  This is the data 
frame that was created.
}
\keyword{datasets}
